Switch event model
Before: the object hooked a function into the global event monitor that
listens for all events of a type and loops through them.
Problems: This doesn't scale (O(n²) is bad for you) and didn't auto-remove
the object afterwards.

Now: Objects are weakly cached. An event is scanned for objects and the
event is sent to those found.